Add user authentication checking in functions

+++ app/Http/Controllers/V1/AlbumController.php
@@ -7,6 +7,7 @@ use App\Models\Album;
 use App\Http\Requests\StoreAlbumRequest;
 use App\Http\Requests\UpdateAlbumRequest;
 use App\Http\Resources\V1\AlbumResource;
+use Illuminate\Support\Facades\Auth;
 
 class AlbumController extends Controller
 {
@@ -15,7 +16,7 @@ class AlbumController extends Controller
      */
     public function index()
     {
-        return AlbumResource::collection(Album::paginate());
+        return AlbumResource::collection(Album::where('user_id', Auth::user()->id)->paginate());
     }
 
     /**
@@ -23,7 +24,9 @@ class AlbumController extends Controller
      */
     public function store(StoreAlbumRequest $request)
     {
-        $album = Album::create($request->all());
+        $data = $request->all();
+        $data['user_id'] = Auth::user()->id;
+        $album = Album::create($data);
         return new AlbumResource($album);
     }
 
@@ -32,6 +35,10 @@ class AlbumController extends Controller
      */
     public function show(Album $album)
     {
+        if(Auth::user()->id != $album->user_id) {
+            return abort(403, "unauthorized");
+        }
+
         return new AlbumResource($album);
     }
 
@@ -40,6 +47,10 @@ class AlbumController extends Controller
      */
     public function update(UpdateAlbumRequest $request, Album $album)
     {
+        if(Auth::user()->id != $album->user_id) {
+            return abort(403, "unauthorized");
+        }
+
         $album->update($request->all());
         return new AlbumResource($album);
     }
@@ -49,6 +60,10 @@ class AlbumController extends Controller
      */
     public function destroy(Album $album)
     {
+        if(Auth::user()->id != $album->user_id) {
+            return abort(403, "unauthorized");
+        }
+        
         $album->delete();
         return response("", 204);
     }
